backend/requirements.txt
uvicorn
fastapi
python-jose
mysql-connector-python